+++ src/utils/random-specific-value-creator.go
@@ -6,62 +6,62 @@ import (
 	common_constants "github.com/pseudoelement/go-file-downloader/src/modules/downloader/constants/common"
 )
 
-func CreateRandomFirstName(min int, max int) string {
+func CreateRandomFirstName(min int64, max int64) string {
 	min, max = setDefaultMinMaxIfZeros(min, max)
 	randomIndex := rand.Intn(len(common_constants.FIRST_NAMES_ARRAY))
 	firstName := common_constants.FIRST_NAMES_ARRAY[randomIndex]
-	if len(firstName) > max || len(firstName) < min {
+	if len(firstName) > int(max) || len(firstName) < int(min) {
 		return CreateRandomFirstName(min, max)
 	} else {
 		return firstName
 	}
 }
 
-func CreateRandomLastName(min int, max int) string {
+func CreateRandomLastName(min int64, max int64) string {
 	min, max = setDefaultMinMaxIfZeros(min, max)
 	randomIndex := rand.Intn(len(common_constants.LAST_NAMES_ARRAY))
 	lastName := common_constants.LAST_NAMES_ARRAY[randomIndex]
-	if len(lastName) > max || len(lastName) < min {
+	if len(lastName) > int(max) || len(lastName) < int(min) {
 		return CreateRandomLastName(min, max)
 	} else {
 		return lastName
 	}
 }
 
-func CreateRandomCarName(min int, max int) string {
+func CreateRandomCarName(min int64, max int64) string {
 	min, max = setDefaultMinMaxIfZeros(min, max)
 	randomIndex := rand.Intn(len(common_constants.CARS_ARRAY))
 	carName := common_constants.CARS_ARRAY[randomIndex]
-	if len(carName) > max || len(carName) < min {
+	if len(carName) > int(max) || len(carName) < int(min) {
 		return CreateRandomCarName(min, max)
 	} else {
 		return carName
 	}
 }
 
-func CreateRandomCountryName(min int, max int) string {
+func CreateRandomCountryName(min int64, max int64) string {
 	min, max = setDefaultMinMaxIfZeros(min, max)
 	randomIndex := rand.Intn(len(common_constants.COUNTRIES_ARRAY))
 	countryName := common_constants.COUNTRIES_ARRAY[randomIndex]
-	if len(countryName) > max || len(countryName) < min {
+	if len(countryName) > int(max) || len(countryName) < int(min) {
 		return CreateRandomCountryName(min, max)
 	} else {
 		return countryName
 	}
 }
 
-func CreateRandomWork(min int, max int) string {
+func CreateRandomWork(min int64, max int64) string {
 	min, max = setDefaultMinMaxIfZeros(min, max)
 	randomIndex := rand.Intn(len(common_constants.WORKS_ARRAY))
 	work := common_constants.WORKS_ARRAY[randomIndex]
-	if len(work) > max || len(work) < min {
+	if len(work) > int(max) || len(work) < int(min) {
 		return CreateRandomWork(min, max)
 	} else {
 		return work
 	}
 }
 
-func setDefaultMinMaxIfZeros(min int, max int) (int, int) {
+func setDefaultMinMaxIfZeros(min int64, max int64) (int64, int64) {
 	if min == 0 {
 		min = 1
 	}
